Introduce Pages
- Not yet done for Overview
- Consider adding something like a "action factory" to unify action creation
