+++ app/src/components/details/MetricDetailsRouter.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import { Route, Routes } from "react-router-dom";
+import { MetricEditPage } from "./edit/MetricEditPage";
+import { MetricViewPage } from "./MetricViewPage";
+
+export const MetricDetailsRouter: React.FC = () => (
+  <Routes>
+    <Route path="/" element={<MetricViewPage />} />
+    <Route path="/edit" element={<MetricEditPage />} />
+  </Routes>
+);
